+++ UtilsTest/Mathematics/Numbers/NumberToStringConverterEngineFixesTests.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Utils.NumberToString;
 
-namespace UtilsTest.Mathematics.Numbers;
+namespace UtilsTest.NumberToString;
 
 /// <summary>
 /// Tests for items 29, 30, 31, 45, 46 of TODO.md (engine robustness and API symmetry fixes).